Ya hice funcionar correctamente CLIENTE ademas ya esta sin estilos

+++ database/migrations/2024_11_24_010529_remove_monto_total_from_reserva_table.php
@@ -13,9 +13,11 @@ class RemoveMontoTotalFromReservaTable extends Migration
      */
     public function up()
     {
-        Schema::table('reserva', function (Blueprint $table) {
-            $table->dropColumn('monto_total');
-        });
+        if (!Schema::hasTable('reserva')) {
+            Schema::table('reserva', function (Blueprint $table) {
+                $table->dropColumn('monto_total');
+            });
+        }
     }
 
     /**
